+++ tests/job/test_abci_job.py
@@ -7,8 +7,8 @@ from unittest.mock import patch
 import pytest
 from utils import qstat_xml
 
-from aiaccel.job import AbciJob
-from aiaccel.job.jobs.job_status import JobStatus
+from aiaccel.hpo.job import AbciJob
+from aiaccel.hpo.job.jobs.job_status import JobStatus
 
 #
