+++ tests/job/test_abci_job_manager.py
@@ -6,8 +6,8 @@ from unittest.mock import MagicMock, patch
 import pytest
 from utils import qstat_xml
 
-from aiaccel.job import AbciJob, AbciJobExecutor
-from aiaccel.job.jobs.job_status import JobStatus
+from aiaccel.hpo.job import AbciJob, AbciJobExecutor
+from aiaccel.hpo.job.jobs.job_status import JobStatus
 
 
 @pytest.fixture
